Fix Deployment Details Bug (#699)
* Improve loading of deployment details

* Hide missing deployment details labels

* Load deployment details once via init()

* Add comment about go loadDeploymentDetails()

* Simplify deployment_details.go

* Remove getDeploymentDetailsIfLoaded() function
